+++ src/Microsoft.AspNetCore.MatrixParameter.Samples/Startup.cs
@@ -1,5 +1,7 @@
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.MatrixParameter.ModelBinders;
+using Microsoft.AspNetCore.MatrixParameter.Swagger;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
