+++ src/Microsoft.AspNetCore.MatrixParameter/ModelBinders/SegmentPrefixAttributeModelBinder.cs
@@ -1,10 +1,11 @@
 using System;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.MatrixParameter.Attributes;
 using Microsoft.AspNetCore.Mvc.ModelBinding;
 using Microsoft.AspNetCore.Mvc.ModelBinding.Metadata;
 
-namespace Microsoft.AspNetCore.MatrixParameter
+namespace Microsoft.AspNetCore.MatrixParameter.ModelBinders
 {
     public class SegmentPrefixAttributeModelBinderProvider : IModelBinderProvider
     {
